# backend/src/nbaforecast/api/routers/games.py
"""Games + predictions router — backend-api.md §3 + Prompt 4."""

from datetime import date as date_type

from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.ext.asyncio import AsyncSession

from nbaforecast.api import services
from nbaforecast.api.deps import get_db_session, get_model_provider
from nbaforecast.api.model_provider import ModelProvider
from nbaforecast.api.schemas.common import Page
from nbaforecast.api.schemas.games import (
    GameBoxScore,
    GameDetail,
    GamePrediction,
    GameSummary,
    GameWinProbabilityTimeline,
)

router = APIRouter(prefix="/games", tags=["games"])


@router.get("", response_model=Page[GameSummary])
async def list_games(
    game_date: date_type | None = None,
    season: str | None = None,
    team: int | None = None,
    page: int = 1,
    page_size: int = 25,
    session: AsyncSession = Depends(get_db_session),
) -> Page[GameSummary]:
    return await services.games.list_games(
        session, game_date=game_date, season=season, team_id=team, page=page, page_size=page_size
    )


@router.get("/{game_id}", response_model=GameDetail)
async def get_game(game_id: str, session: AsyncSession = Depends(get_db_session)) -> GameDetail:
    game = await services.games.get_game(session, game_id)
    if game is None:
        raise HTTPException(status_code=404, detail=f"game {game_id!r} not found")
    return game


@router.get("/{game_id}/boxscore", response_model=GameBoxScore)
async def get_game_boxscore(
    game_id: str, session: AsyncSession = Depends(get_db_session)
) -> GameBoxScore:
    boxscore = await services.games.get_game_boxscore(session, game_id)
    if boxscore is None:
        raise HTTPException(status_code=404, detail=f"no box score available for game {game_id!r}")
    return boxscore


async def _predict_or_404(
    session: AsyncSession, model_provider: ModelProvider, game_id: str, *, full: bool
) -> GamePrediction:
    try:
        prediction = await services.games.get_game_prediction(
            session, model_provider, game_id, full=full
        )
    except RuntimeError as exc:
        # ModelProvider.get() raises this when no champion has been loaded yet — a transient
        # startup/MLflow-outage state, not a client error.
        raise HTTPException(status_code=503, detail=str(exc)) from exc
    if prediction is None:
        raise HTTPException(status_code=404, detail=f"no prediction available for game {game_id!r}")
    return prediction


@router.get("/{game_id}/win-probability", response_model=GameWinProbabilityTimeline)
async def get_game_win_probability(
    game_id: str,
    session: AsyncSession = Depends(get_db_session),
    model_provider: ModelProvider = Depends(get_model_provider),
) -> GameWinProbabilityTimeline:
    try:
        timeline = await services.games.game_win_probability(session, model_provider, game_id)
    except RuntimeError as exc:
        # No in-game champion loaded yet (transient MLflow/startup) — mirror the prediction 503.
        raise HTTPException(status_code=503, detail=str(exc)) from exc
    if timeline is None:
        raise HTTPException(
            status_code=404, detail=f"no win-probability timeline for game {game_id!r}"
        )
    return timeline


@router.get("/{game_id}/prediction", response_model=GamePrediction)
async def get_game_prediction(
    game_id: str,
    session: AsyncSession = Depends(get_db_session),
    model_provider: ModelProvider = Depends(get_model_provider),
) -> GamePrediction:
    return await _predict_or_404(session, model_provider, game_id, full=False)


@router.get("/{game_id}/prediction/full-explanation", response_model=GamePrediction)
async def get_game_prediction_full_explanation(
    game_id: str,
    session: AsyncSession = Depends(get_db_session),
    model_provider: ModelProvider = Depends(get_model_provider),
) -> GamePrediction:
    return await _predict_or_404(session, model_provider, game_id, full=True)