# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "NBAforecast",
  "description": "Explainable NBA predictions, stats hub, and live win probability.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/games": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "List Games",
    "operationId": "list_games_api_v1_games_get",
    "parameters": [
     {
      "name": "game_date",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Game Date"
      }
     },
     {
      "name": "season",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Season"
      }
     },
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Team"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_GameSummary_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/games/{game_id}": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "Get Game",
    "operationId": "get_game_api_v1_games__game_id__get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/games/{game_id}/boxscore": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "Get Game Boxscore",
    "operationId": "get_game_boxscore_api_v1_games__game_id__boxscore_get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameBoxScore"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/games/{game_id}/win-probability": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "Get Game Win Probability",
    "operationId": "get_game_win_probability_api_v1_games__game_id__win_probability_get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameWinProbabilityTimeline"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/games/{game_id}/prediction": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "Get Game Prediction",
    "operationId": "get_game_prediction_api_v1_games__game_id__prediction_get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GamePrediction"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/games/{game_id}/prediction/full-explanation": {
   "get": {
    "tags": [
     "games"
    ],
    "summary": "Get Game Prediction Full Explanation",
    "operationId": "get_game_prediction_full_explanation_api_v1_games__game_id__prediction_full_explanation_get",
    "parameters": [
     {
      "name": "game_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Game Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GamePrediction"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "BoxScorePlayerLine": {
    "properties": {
     "player_id": {
      "type": "integer",
      "title": "Player Id"
     },
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "started": {
      "type": "boolean",
      "title": "Started"
     },
     "min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min"
     },
     "pts": {
      "type": "integer",
      "title": "Pts"
     },
     "reb": {
      "type": "integer",
      "title": "Reb"
     },
     "ast": {
      "type": "integer",
      "title": "Ast"
     },
     "stl": {
      "type": "integer",
      "title": "Stl"
     },
     "blk": {
      "type": "integer",
      "title": "Blk"
     },
     "tov": {
      "type": "integer",
      "title": "Tov"
     },
     "fgm": {
      "type": "integer",
      "title": "Fgm"
     },
     "fga": {
      "type": "integer",
      "title": "Fga"
     },
     "fg3m": {
      "type": "integer",
      "title": "Fg3M"
     },
     "fg3a": {
      "type": "integer",
      "title": "Fg3A"
     },
     "ftm": {
      "type": "integer",
      "title": "Ftm"
     },
     "fta": {
      "type": "integer",
      "title": "Fta"
     },
     "plus_minus": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plus Minus"
     }
    },
    "type": "object",
    "required": [
     "player_id",
     "full_name",
     "started",
     "min",
     "pts",
     "reb",
     "ast",
     "stl",
     "blk",
     "tov",
     "fgm",
     "fga",
     "fg3m",
     "fg3a",
     "ftm",
     "fta",
     "plus_minus"
    ],
    "title": "BoxScorePlayerLine",
    "description": "One player's box-score line within a game box score."
   },
   "BoxScoreTeam": {
    "properties": {
     "team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "is_home": {
      "type": "boolean",
      "title": "Is Home"
     },
     "pts": {
      "type": "integer",
      "title": "Pts"
     },
     "reb": {
      "type": "integer",
      "title": "Reb"
     },
     "ast": {
      "type": "integer",
      "title": "Ast"
     },
     "stl": {
      "type": "integer",
      "title": "Stl"
     },
     "blk": {
      "type": "integer",
      "title": "Blk"
     },
     "tov": {
      "type": "integer",
      "title": "Tov"
     },
     "fgm": {
      "type": "integer",
      "title": "Fgm"
     },
     "fga": {
      "type": "integer",
      "title": "Fga"
     },
     "fg3m": {
      "type": "integer",
      "title": "Fg3M"
     },
     "fg3a": {
      "type": "integer",
      "title": "Fg3A"
     },
     "ftm": {
      "type": "integer",
      "title": "Ftm"
     },
     "fta": {
      "type": "integer",
      "title": "Fta"
     },
     "players": {
      "items": {
       "$ref": "#/components/schemas/BoxScorePlayerLine"
      },
      "type": "array",
      "title": "Players"
     }
    },
    "type": "object",
    "required": [
     "team",
     "is_home",
     "pts",
     "reb",
     "ast",
     "stl",
     "blk",
     "tov",
     "fgm",
     "fga",
     "fg3m",
     "fg3a",
     "ftm",
     "fta",
     "players"
    ],
    "title": "BoxScoreTeam",
    "description": "One team's totals plus its player lines, ordered starters-then-minutes."
   },
   "Contribution": {
    "properties": {
     "feature": {
      "type": "string",
      "title": "Feature"
     },
     "display_label": {
      "type": "string",
      "title": "Display Label"
     },
     "raw_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "integer"
       },
       {
        "type": "string"
       },
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Raw Value"
     },
     "formatted_value": {
      "type": "string",
      "title": "Formatted Value"
     },
     "contribution": {
      "type": "number",
      "title": "Contribution"
     },
     "direction": {
      "type": "string",
      "enum": [
       "up",
       "down"
      ],
      "title": "Direction"
     }
    },
    "type": "object",
    "required": [
     "feature",
     "display_label",
     "raw_value",
     "formatted_value",
     "contribution",
     "direction"
    ],
    "title": "Contribution",
    "description": "One feature's signed contribution to a single prediction.\n\n``display_label``/``formatted_value`` are placeholders (the raw feature name and\n``str(raw_value)``) until T2.11's humanizer decorates them \u2014 see\n``explain/explainers.py``'s module docstring."
   },
   "Explanation": {
    "properties": {
     "baseline": {
      "type": "number",
      "title": "Baseline"
     },
     "prediction": {
      "type": "number",
      "title": "Prediction"
     },
     "contributions": {
      "items": {
       "$ref": "#/components/schemas/Contribution"
      },
      "type": "array",
      "title": "Contributions"
     },
     "units": {
      "$ref": "#/components/schemas/ExplanationUnits"
     },
     "notes": {
      "type": "string",
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "baseline",
     "prediction",
     "contributions",
     "units",
     "notes"
    ],
    "title": "Explanation",
    "description": "The \u00a75 explanation contract \u2014 every model head's ``explain()`` returns this shape.\n\n``contributions`` is sorted by ``abs(contribution)`` descending (\u00a75: \"sorted by magnitude\")\nso the frontend can show the top-N headline drivers with the rest expandable. Additivity\nholds in whichever ``units`` the explanation was computed in:\n``sum(c.contribution for c in contributions) == prediction - baseline`` (T2.12 verifies this\nfor the SHAP path within floating-point tolerance)."
   },
   "ExplanationUnits": {
    "type": "string",
    "enum": [
     "probability_points",
     "log_odds",
     "points",
     "rebounds",
     "assists",
     "three_pointers_made"
    ],
    "title": "ExplanationUnits",
    "description": "What scale ``Explanation.baseline``/``prediction``/``Contribution.contribution`` are in."
   },
   "GameBoxScore": {
    "properties": {
     "game_id": {
      "type": "string",
      "title": "Game Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "home": {
      "$ref": "#/components/schemas/BoxScoreTeam"
     },
     "away": {
      "$ref": "#/components/schemas/BoxScoreTeam"
     }
    },
    "type": "object",
    "required": [
     "game_id",
     "status",
     "home",
     "away"
    ],
    "title": "GameBoxScore",
    "description": "``GET /games/{game_id}/boxscore`` \u2014 both teams' totals and player lines for a played game."
   },
   "GameDetail": {
    "properties": {
     "game_id": {
      "type": "string",
      "title": "Game Id"
     },
     "season": {
      "type": "string",
      "title": "Season"
     },
     "game_date": {
      "type": "string",
      "format": "date",
      "title": "Game Date"
     },
     "home_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "away_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "home_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home Score"
     },
     "away_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Away Score"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "game_datetime": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Game Datetime"
     },
     "num_periods": {
      "type": "integer",
      "title": "Num Periods"
     }
    },
    "type": "object",
    "required": [
     "game_id",
     "season",
     "game_date",
     "home_team",
     "away_team",
     "home_score",
     "away_score",
     "status",
     "game_datetime",
     "num_periods"
    ],
    "title": "GameDetail",
    "description": "Full game record \u2014 backend-api.md \u00a73 ``GET /games/{game_id}``."
   },
   "GamePrediction": {
    "properties": {
     "game_id": {
      "type": "string",
      "title": "Game Id"
     },
     "win_prob": {
      "type": "number",
      "title": "Win Prob"
     },
     "margin": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Margin"
     },
     "total": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     },
     "market": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Market"
     },
     "explanation": {
      "$ref": "#/components/schemas/Explanation"
     }
    },
    "type": "object",
    "required": [
     "game_id",
     "win_prob",
     "explanation"
    ],
    "title": "GamePrediction",
    "description": "backend-api.md \u00a74 ``GamePrediction``.\n\n``margin``/``total`` are reserved for the regressor heads (T3.1, M3 scope) \u2014 ``None`` until\nthen. ``market`` is reserved for the v2 odds-comparison feature (modeling.md \u00a710)."
   },
   "GameSummary": {
    "properties": {
     "game_id": {
      "type": "string",
      "title": "Game Id"
     },
     "season": {
      "type": "string",
      "title": "Season"
     },
     "game_date": {
      "type": "string",
      "format": "date",
      "title": "Game Date"
     },
     "home_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "away_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "home_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home Score"
     },
     "away_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Away Score"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "game_id",
     "season",
     "game_date",
     "home_team",
     "away_team",
     "home_score",
     "away_score",
     "status"
    ],
    "title": "GameSummary",
    "description": "One row of a games list \u2014 backend-api.md \u00a73 ``GET /games``."
   },
   "GameWinProbabilityTimeline": {
    "properties": {
     "game_id": {
      "type": "string",
      "title": "Game Id"
     },
     "home_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "away_team": {
      "$ref": "#/components/schemas/nbaforecast__api__schemas__games__TeamSummary"
     },
     "points": {
      "items": {
       "$ref": "#/components/schemas/WinProbabilityPoint"
      },
      "type": "array",
      "title": "Points"
     }
    },
    "type": "object",
    "required": [
     "game_id",
     "home_team",
     "away_team",
     "points"
    ],
    "title": "GameWinProbabilityTimeline",
    "description": "``GET /games/{game_id}/win-probability`` \u2014 the in-game win-prob trajectory for a game.\n\nReplays the stored play-by-play through the in-game win-prob champion: one point per usable\nevent, each with the home win probability and the SHAP drivers behind it at that moment."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Page_GameSummary_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/GameSummary"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[GameSummary]"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WinProbabilityDriver": {
    "properties": {
     "label": {
      "type": "string",
      "title": "Label"
     },
     "value": {
      "type": "string",
      "title": "Value"
     },
     "contribution": {
      "type": "number",
      "title": "Contribution"
     },
     "direction": {
      "type": "string",
      "title": "Direction"
     }
    },
    "type": "object",
    "required": [
     "label",
     "value",
     "contribution",
     "direction"
    ],
    "title": "WinProbabilityDriver",
    "description": "One factor moving the in-game win probability at a moment (thinking panel)."
   },
   "WinProbabilityPoint": {
    "properties": {
     "event_num": {
      "type": "integer",
      "title": "Event Num"
     },
     "period": {
      "type": "integer",
      "title": "Period"
     },
     "clock": {
      "type": "string",
      "title": "Clock"
     },
     "seconds_remaining": {
      "type": "integer",
      "title": "Seconds Remaining"
     },
     "home_score": {
      "type": "integer",
      "title": "Home Score"
     },
     "away_score": {
      "type": "integer",
      "title": "Away Score"
     },
     "home_win_prob": {
      "type": "number",
      "title": "Home Win Prob"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "drivers": {
      "items": {
       "$ref": "#/components/schemas/WinProbabilityDriver"
      },
      "type": "array",
      "title": "Drivers"
     }
    },
    "type": "object",
    "required": [
     "event_num",
     "period",
     "clock",
     "seconds_remaining",
     "home_score",
     "away_score",
     "home_win_prob",
     "description",
     "drivers"
    ],
    "title": "WinProbabilityPoint",
    "description": "One moment on the in-game win-probability timeline (one play-by-play event)."
   },
   "nbaforecast__api__schemas__games__TeamSummary": {
    "properties": {
     "team_id": {
      "type": "integer",
      "title": "Team Id"
     },
     "abbreviation": {
      "type": "string",
      "title": "Abbreviation"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     }
    },
    "type": "object",
    "required": [
     "team_id",
     "abbreviation",
     "full_name"
    ],
    "title": "TeamSummary",
    "description": "The minimal team identity embedded in a game response."
   },
...
  }
 }
}